feat(printing): Move print() to imageBase
Moving the print function to imageBase will allow
MultiImageViewer to inherit the functionality.
