+++ src/lib/viewers/image/__tests__/ImageBaseViewer-test.html
@@ -1,13 +1,18 @@
 <style>
-    html, body {
+    html,
+    body {
         width: 100%;
         height: 100%;
     }
     .container {
         width: 500px;
         height: 500px;
-        margin: 200px; 
+        margin: 200px;
         background-color: #eee;
     }
 </style>
-<div class="container"><div class="bp-container"><div class="bp"></div></div></div>
+<div class="container">
+    <div class="bp-container">
+        <div class="bp"><img /></div>
+    </div>
+</div>
